fix: avoid circular dependencies
