+++ src/app/drive/services/worker.service/downloadWorkerHandler.test.ts
@@ -4,10 +4,11 @@ vi.mock('../download.service/downloadFileFromBlob', () => ({
 
 import streamSaver from 'streamsaver';
 import { describe, test, expect, vi, Mock, beforeEach } from 'vitest';
-import { DownloadAbortedByUserError, downloadWorkerHandler } from './downloadWorkerHandler';
+import { downloadWorkerHandler } from './downloadWorkerHandler';
 import { DriveFileData } from 'app/drive/types';
 import { MockWorker } from '../../../../__mocks__/WebWorker';
 import downloadFileFromBlob from '../download.service/downloadFileFromBlob';
+import { DownloadAbortedByUserError } from 'app/network/errors/download.errors';
 
 const writeMock = vi.fn();
 const closeMock = vi.fn();
